+++ src/main/kotlin/me/rivia/api/database/entry/Usage.kt
@@ -3,6 +3,6 @@ package me.rivia.api.database.entry
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean
 
 @DynamoDbBean
-data class Usage(var usageId: String?, var tenantId: String?, var userId: String?, var timings: List<Float>?) {
+data class Usage(var usageId: String?, var tenantId: String?, var userId: String?, var timings: List<Double>?) {
     constructor() : this(null, null, null, null)
 }
